backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
Flask-CORS==3.0.10
PyMySQL==1.0.2
python-dotenv==0.19.0
Flask-JWT-Extended==4.3.1
werkzeug==2.0.3
requests==2.31.0
Pillow==10.3.0
